% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftoolsPipableCmd.R
\name{bcftoolsPipableCmd}
\alias{bcftoolsPipableCmd}
\title{bcftoolsPipableCmd}
\usage{
bcftoolsPipableCmd(
  files,
  output_name = NULL,
  output_type = NULL,
  regions = NULL,
  variant_ids = NULL,
  sample_ids = NULL,
  query = NULL,
  chr_nm_map = NULL,
  exclude_annos = NULL,
  extra_cmds = NULL,
  scratch_dir = "/tmp",
  bcftools_bin = "bcftools"
)
}
\arguments{
\item{files}{Character vector of VCF/BCF file paths and/or URLs. If multiple files, they will be concatenated, but must have the same sample columns. Combination will be much faster if no regions are given to subset by, the files are either all VCF or all BCF, and all their headers are compatible (useful to combine VCFs split by chromosome).}

\item{output_name}{Filename of output file. Filename suffix determines the type (.vcf, .vcf.bgz, .bcf, .bcf.bgz).}

\item{output_type}{"z"/"v" = compressed/uncompressed VCF, "b"/"u" = compressed/uncompressed BCF. Note that the file extension of output_name will take precedence.}

\item{regions}{data.frame of bed-like format (1st column chromosome, 2nd start position (0-indexed), 3rd end, other cols unused)}

\item{variant_ids}{Character vector of variant IDs, to subset the VCF/BCF(s) by.}

\item{sample_ids}{Character vector of sample IDs, to subset the VCF/BCF(s) by.}

\item{query}{<TODO document me>}

\item{chr_nm_map}{data.frame with two columns. Used to convert the VCF files' chromosome/contig names from the names in the first column to the second. E.g. to convert from "chr1" to "1". You can assume this name mapping is applied before the VCF is subset by regions and/or variant_ids.}

\item{exclude_annos}{<Document TODO, maybe kinda redundant w/ info_fields and format_fields options? see docs of bcftools anno -x>}

\item{extra_cmds}{<Document TODO, I know you've been thinking about rm'ing this but I think it's useful, since o/w would have to write to an intermediate file; the advantage of tacking on cmds like this is you can create one big single ~pipable~ cmd>}

\item{scratch_dir}{Directory where temporary files should be written.}

\item{bcftools_bin}{Path to the program "bcftools".}

\item{info_fields}{<TODO> Character vector of INFO fields to keep, e.g. c("AC","AF","AN"). If left NULL, all fields are gotten.}

\item{format_fields}{<TODO> Character vector of FORMAT fields to keep, e.g. c("GT","DP"). If left NULL, all fields are gotten.}
}
\description{
A wrapper for a couple parts of bcftools. Combine and subset VCF/BCF files.
}
\section{A tip to increase speed}{
 Even if you already provide variant_ids, also providing regions will speed things up a lot. It is easier for computers to compare chromosome and position numbers than ID strings.
}

\section{TODO}{
 Support and chr:pos(:ref:alt) IDs? Support mixed types of IDs? Infer regions if all IDs are chr:pos(:ref:alt), and maybe mention in these docs that that optimization is made once implemented. Allow GRanges regions input?
}

