% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldCalc.R
\name{ldCalc}
\alias{ldCalc}
\title{ldCalc}
\usage{
ldCalc(
  gds,
  regions = NULL,
  variant_ids = NULL,
  sample_ids = NULL,
  ref_alleles = NULL,
  maf_thresh = NaN,
  mac_thresh = 1L,
  variant_missingness_thresh = NaN,
  slide = 0,
  method = "corr",
  n_threads = parallel::detectCores(),
  yes_really = F
)
}
\arguments{
\item{gds}{A `SeqArray` GDS file handle (obtained from `seqOpen("example.gds")`)}

\item{regions}{Character vector of regions of the format chr:start-end, or a data.frame of bed-like format (1st column chromosome, 2nd start position (0-indexed), 3rd end, other cols unused)}

\item{variant_ids}{Character vector of variant IDs. If given, the resulting LD matrix's colnames will be variant IDs. Otherwise, the colnames will be of the format chr:pos:ref:alt.}

\item{sample_ids}{Character vector of sample IDs, to subset the VCF/BCF(s) by.}

\item{ref_alleles}{Character vector of variant reference alleles corresponding to variant_ids.}

\item{maf_thresh}{MAF threshold}

\item{mac_thresh}{MAC threshold}

\item{slide}{Size of sliding window. Full LD matrix will be computed by default.}

\item{method}{Measure of LD to use. "corr" by default, referring to Pearson r. See `?snpgdsLDMat` from the `SNPRelate` package for more details.}

\item{n_threads}{Number of threads for parallel processing.}

\item{yes_really}{Safety measure. LD for a large number of variants can consume a lot of hard drive space. If calculating LD for a very large number of variants (>20k), pass yes_really=T to confirm you wish to do this.}

\item{variant_missingness}{Missingness threshold}
}
\description{
Calculate LD from `SeqArray` GDS files.
}
