% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftOver.R
\name{liftOver}
\alias{liftOver}
\title{liftOver}
\usage{
liftOver(
  to_lift,
  from_build,
  to_build,
  out_file = NULL,
  liftover_bin = "liftOver"
)
}
\arguments{
\item{to_lift}{A BED format file, or a BED-like data.frame. NOTE: UCSC-style chromosome naming must be used (e.g. "chr1", not just "1"), or liftOver won't work.}

\item{from_build}{A string, e.g. "hg19"}

\item{to_build}{A string, e.g. "hg38"}

\item{out_file}{Desired output filename. Results are also returned as an R data.frame whether this option is given or not.}

\item{liftover_bin}{Path to the program liftOver. Can be downloaded from http://hgdownload.soe.ucsc.edu/admin/exe/linux.x86_64/liftOver}
}
\value{
A BED-like data frame with lifted genome coordinates.
}
\description{
Simple wrapper for UCSC's liftOver, to convert genome coordinates between reference genomes. Required chain files are automatically downloaded. BED file specification: http://www.genome.ucsc.edu/FAQ/FAQformat.html#format1
}
\section{Command-line use}{
 Since this function can accept files, you could run it from the command line like so: R -q -e source("liftOver.R"); liftOver("my_file.bed", "hg19", "hg38", liftover_bin="path/to/liftOver"). At that point you should probably just use liftOver by itself though.
}

