% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfs2Gds.R
\name{vcfs2Gds}
\alias{vcfs2Gds}
\title{Vcfs2Gds}
\usage{
vcfs2Gds(
  files,
  output_name,
  regions = NULL,
  variant_ids = NULL,
  sample_ids = NULL,
  chr_nm_map = NULL,
  exclude_annos = NULL,
  extra_cmds = NULL,
  scratch_dir = "/tmp/",
  bcftools_bin = "bcftools"
)
}
\arguments{
\item{files}{VCF file path(s) or URL(s).}

\item{output_name}{Filename of output GDS file.}

\item{regions}{A data.frame of bed-like format (1st column chrom, 2nd chromStart, 3rd chromEnd, other cols unused)}

\item{variant_ids}{Character vector of variant IDs, to subset the VCFs by.}

\item{sample_ids}{Character vector of sample IDs, to subset the VCFs by.}

\item{chr_nm_map}{A data.frame with two columns. Used to convert the VCF files' chromosome/contig names from the names in the first column to the second. E.g. to convert from "chr1" to "1".}

\item{exclude_annos}{<TODO, see bcftoolsPipableCmd(), maybe redundant w/ info_fields and format_fields params, choose one>}

\item{extra_cmds}{<TODO, document>}

\item{bcftools_bin}{Path to the program "bcftools".}

\item{storage.option}{<TODO, not implemented> Storage/Compression options. Passed directly to seqVCF2GDS. See ?seqVCF2GDS and ?seqStorageOption for details.}
}
\description{
Convert a bunch of VCF files into one SeqArray GDS file.
}
\section{A tip to increase speed}{
 Even if you already provide variant_ids, also providing regions will speed things up a lot. It is easier for computers to compare chromosome and position numbers than ID strings.
}

\section{TODO}{
 Support and chr:pos(:ref:alt) IDs? Support mixed types of IDs? Infer regions if all IDs are chr:pos(:ref:alt), and maybe mention in these docs that that optimization is made once implemented. Deal with "chr#" vs. "#" syntax for the user of not matching (I think bcftools norm or  bcftools annotate --rename-chrs can help). Allow GRanges regions input? Support BCF files too?--Note would have to use pipe(..., "rb") maybe.
}

